requirements.txt
bleach (1.5.0)
certifi (2017.11.5)
chardet (3.0.4)
Django (2.0)
enum34 (1.1.6)
html5lib (0.9999999)
idna (2.6)
image (1.5.17)
Keras (2.1.2)
Markdown (2.6.10)
numpy (1.13.3)
olefile (0.44)
pandas (0.21.1)
Pillow (4.3.0)
pip (9.0.1)
protobuf (3.5.1)
python-dateutil (2.6.1)
pytz (2017.3)
PyYAML (3.12)
requests (2.18.4)
scipy (1.0.0)
setuptools (38.2.5)
six (1.11.0)
tensorflow (1.4.1)
tensorflow-tensorboard (0.4.0rc3)
urllib3 (1.22)
Werkzeug (0.13)
wheel (0.30.0)
